[Silver III] Title: 2×n 타일링, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/11726. 2×n 타일링/2×n 타일링.cc
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-long long dp[1001];
-long long  Func(int n) // 직사각형을 채우는 방법의 수를 리턴할 함수
+int dp[1001];
+int  Func(int n) // 직사각형을 채우는 방법의 수를 리턴할 함수
 {
 	if (n == 1) return 1; // 2x1 직사각형의 경우 1개의 방법만 존재
 	if (n == 2) return 2; // 2x2 직사각형의 경우 2개의 방법만 존재
